backend/requirements.txt
crewai>=0.80.0
langchain-groq>=0.2.0
requests>=2.31.0
supabase>=2.4.0
python-dotenv>=1.0.0
